Major refactor at core
In order to abstract the commands creation into something that people
are used to, this commit changes the bot core to a middleware
abstraction like Ring and Compojure. The goal is keep the core
agnostic to the bot platform, that said you won't find Slack functions
inside of it.

Now the bot consists of four components:

**Handlers** are functions that define your bot behavior in a imperative
way.

**Messages** are Clojure maps with standard keys.

**Reply** are also Clojure maps with standard keys which can be translated
to a valid text message downstream.

**Middlewares** are higher-level functions that add additional
functionality to handlers.

Slack processor:

The Slack processor component is responsible for connect the RTM API
to the handlers, transforming the event into a message, passing it
to a pipeline of commands, produce a valid reply, and send it back to
the open socket.

Message and reply are being enhanced with Slack channels in order to
direct the message to the right user. At the moment, the relation
between input and output channel is 1-1, but it could change in the
future to even support multiple output channels.

Things to be done and improved:
- Support commands with arguments
- Support async handlers (long-term)

Reference:
- https://github.com/weavejester/compojure
- https://github.com/ring-clojure/ring